requirements.txt
Babel (2.3.4)
blinker (1.4)
chardet (2.3.0)
click (6.7)
cssselect (1.0.1)
decorator (4.0.11)
Django (1.10.5)~
django-angular (0.8.4)~
django-classy-tags (0.8.0)~
django-filter (1.0.1)~
django-sekizai (0.10.0)~
djangorestframework (3.5.3)~
docutils (0.13.1)
Flask (0.12)
Flask-Babel (0.11.1)
Flask-Login (0.4.0)
Flask-Mail (0.9.1)
Flask-OpenID (1.2.5)
Flask-SQLAlchemy (2.1)
Flask-WhooshAlchemy (0.56)
Flask-WTF (0.14.2)
flup (1.0.2)~
funcsigs (1.0.2)
itsdangerous (0.24)
Jinja2 (2.9.4)
lamson (1.3.4)
lockfile (0.12.2)
Markdown (2.6.7)~
MarkupSafe (0.23)~
mock (2.0.0)
nose (1.3.7)
pbr (1.10.0)
pip (9.0.1)
Pygments (2.2.0)
python-daemon (2.1.2)
python-modargs (1.7)
python-openid (2.2.5)
pytz (2016.10)
setuptools (33.1.1)
six (1.10.0)
SQLAlchemy (1.1.5)
sqlalchemy-migrate (0.10.0)
sqlparse (0.2.2)
Tempita (0.5.2)
Werkzeug (0.11.15)
Whoosh (2.7.4)
WTForms (2.1)